+++ hc/api/tests/test_get_check.py
@@ -1,5 +1,4 @@
 from datetime import timedelta as td
-import uuid
 
 from django.utils.timezone import now
 from hc.api.models import Channel, Check